requirements.txt
google-adk
litellm
playwright
playwright-stealth
python-dotenv
pdfplumber
google-genai
